+++ src/components/common/modal/useModal.ts
@@ -1,13 +1,35 @@
-import { useState } from "react";
+import { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import useLogin from "../../../hooks/useLogin";
 import useSignup from "../../../hooks/useSignup";
+
+interface UpdateParams {
+  key: "authenticationMethod" | "authenticationContent";
+  value: string;
+}
 
 const useModal = ({ onClose }: { onClose: () => void }) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [displayName, setDisplayName] = useState("");
   const [EntranceCode, setEntranceCode] = useState("");
+
+  const inputRef = useRef(null);
+  const [isShowImage, setIsShowImage] = useState(false);
+  const [newDone, setNewDone] = useState({
+    complitedAt: new Date(),
+    authenticationMethod: "",
+    authenticationContent: "",
+  });
+
+  const handleUpdate = (params: UpdateParams[]) => {
+    setNewDone({
+      ...newDone,
+      [params[0].key]: params[0].value,
+      [params[1].key]: params[1].value,
+    });
+  };
+
   const { signup } = useSignup();
   const { login } = useLogin();
   const navigate = useNavigate();
@@ -47,14 +69,25 @@ const useModal = ({ onClose }: { onClose: () => void }) => {
     navigate("/room");
   };
 
+  const handleCertificationSubmit = (event: any) => {
+    event.preventDefault();
+    //submit 처리 로직 db로 전송
+  };
+
   return {
     handleData,
     handleSignupSubmit,
     handleLoginSubmit,
     handleCodeSubmit,
+    handleUpdate,
     email,
     password,
     displayName,
+    inputRef,
+    isShowImage,
+    setIsShowImage,
+    newDone,
+    handleCertificationSubmit,
   };
 };
